v1.2
Added new annotation params:
1. firstRowColumnNames()
2. rowDataDelimiter()

Supports w/ and w/o column names in the CSV file. If column names doesn't exist in the first row of the CSV file, specify in the @TestData annotation as
@TestData(file="HomePageVerticalLinks.csv", firstRowColumnNames = false)

Also, can specify a custom delimiter for row data. If so, specify like below:
@TestData(file="HomePageHorizontalLinks.csv", rowDataDelimiter = ";")

Both can be combined as well.